# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "onec",
    "version": "2.0.0"
  },
...
  "paths": {
...
    "/contract": {
      "post": {
        "operationId": "createContract",
        "tags": [
          "contract"
        ],
        "description": "Create a new Datacontract",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "nullable": true
                  },
                  "client": {
                    "type": "string",
                    "nullable": true
                  },
                  "state": {
                    "type": "string",
                    "nullable": true
                  },
                  "cnpj": {
                    "type": "string",
                    "nullable": true
                  },
                  "sindic": {
                    "type": "string",
                    "nullable": true
                  },
                  "year": {
                    "type": "string",
                    "nullable": true
                  },
                  "matter": {
                    "type": "string",
                    "nullable": true
                  },
                  "forecast": {
                    "type": "string",
                    "nullable": true
                  },
                  "contractTotal": {
                    "type": "string",
                    "nullable": true
                  },
                  "percentage": {
                    "type": "number",
                    "nullable": true
                  },
                  "signedContract": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "nullable": true
                  },
                  "averageGuide": {
                    "type": "number",
                    "nullable": true
                  },
                  "partner": {
                    "type": "string",
                    "nullable": true
                  },
                  "partnerCommission": {
                    "type": "number",
                    "nullable": true
                  },
                  "counter": {
                    "type": "string",
                    "nullable": true
                  },
                  "email": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "city",
                  "client",
                  "state",
                  "cnpj",
                  "sindic",
                  "year",
                  "matter",
                  "forecast",
                  "contractTotal",
                  "percentage",
                  "signedContract",
                  "status",
                  "averageGuide",
                  "partner",
                  "partnerCommission",
                  "counter",
                  "email"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getContract",
        "tags": [
          "contract"
        ],
        "description": "Get a list of contract",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string",
                        "nullable": true
                      },
                      "client": {
                        "type": "string",
                        "nullable": true
                      },
                      "state": {
                        "type": "string",
                        "nullable": true
                      },
                      "cnpj": {
                        "type": "string",
                        "nullable": true
                      },
                      "sindic": {
                        "type": "string",
                        "nullable": true
                      },
                      "year": {
                        "type": "string",
                        "nullable": true
                      },
                      "matter": {
                        "type": "string",
                        "nullable": true
                      },
                      "forecast": {
                        "type": "string",
                        "nullable": true
                      },
                      "contractTotal": {
                        "type": "string",
                        "nullable": true
                      },
                      "percentage": {
                        "type": "number",
                        "nullable": true
                      },
                      "signedContract": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string",
                        "nullable": true
                      },
                      "averageGuide": {
                        "type": "number",
                        "nullable": true
                      },
                      "partner": {
                        "type": "string",
                        "nullable": true
                      },
                      "partnerCommission": {
                        "type": "number",
                        "nullable": true
                      },
                      "counter": {
                        "type": "string",
                        "nullable": true
                      },
                      "email": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "nullable": true
                      },
                      "updatedAt": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "city",
                      "client",
                      "state",
                      "cnpj",
                      "sindic",
                      "year",
                      "matter",
                      "forecast",
                      "contractTotal",
                      "percentage",
                      "signedContract",
                      "status",
                      "averageGuide",
                      "partner",
                      "partnerCommission",
                      "counter",
                      "email",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "operationId": "authenticateUser",
        "tags": [
          "Authentication"
        ],
        "description": "Realiza o login do usuário e retorna o token JWT com validade de 60 dias",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "token"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "operationId": "getProfileUser",
        "tags": [
          "Users"
        ],
        "description": "List all Users",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      "role": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "email",
                      "role"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "operationId": "updateUser",
        "tags": [
          "Users"
        ],
        "description": "Update an existing User",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "tags": [
          "Users"
        ],
        "description": "Delete an existing User",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}